Alternate move axes
Previously the movement on the X axis was completed before starting the
Y axis movement. This could be problematic in situations where the
target coordinate was continually changing because that would cause only
movement on the X axis to occur and also results in less natural looking
movement. By alternating the axis of movement the move will be made to
target coordinate in both axes even if it's never reached. The 45 degree
movement with horizontal/vertical movement at the end is an aesthetic
improvement, though still not ideal. Unfortunately this causes an
increase of 24 bytes program memory and 1 byte SRAM due to the axis
switching code.